requirements.txt
django
djangorestframework
django-cors-headers
whitenoise
django-environ